Applied clang-tidy fixes and added report analysis

+++ src/app/util/MainMenu.cpp
@@ -2,7 +2,7 @@
 
 #include <imgui.h>
 
-void MainMenu::goToScene(unsigned int sceneIndex) {
+void MainMenu::goToScene(unsigned long sceneIndex) {
     *currentSceneIndex = sceneIndex;
     active = false;
     scenes[*currentSceneIndex]->setup();
@@ -11,7 +11,7 @@ void MainMenu::goToScene(unsigned int sceneIndex) {
 void MainMenu::render() {
     ImGui::Begin("Main Menu");
 
-    for (unsigned int i = 0; i < scenes.size(); i++) {
+    for (unsigned long i = 0; i < scenes.size(); i++) {
         const char *btnName = scenes[i]->getName();
         if (ImGui::Button(btnName)) {
             goToScene(i);
